backend/requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
scikit-learn